Split MCPOAuthRoutes.swift along its endpoint seams (#1122)
The 933-line OAuth 2.1 authorization-server file becomes the route
table (MCPOAuthRoutes.swift, 55 lines) plus one file per concern:
the authorize/consent flow (+Authorize), dynamic client registration
(+Registration), the token endpoint with code exchange and refresh
rotation (+Token), the discovery/metadata surface (MCPOAuthSurface),
and the wire DTOs (MCPOAuthWireTypes).

No behavior change: endpoint set, error bodies, and token semantics
are unchanged.
